updating blogs controllers

+++ part4/exercises-blog-list/controllers/blogs.js
@@ -1,4 +1,5 @@
 const blogsRouter = require("express").Router();
+const userExtractor = require("../middlewares/userExtractor");
 const Blog = require("../models/Blog");
 const User = require("../models/User");
 
@@ -30,16 +31,17 @@ blogsRouter.get("/:id", async (request, response, next) => {
   }
 });
 
-blogsRouter.post("/", async (request, response, next) => {
-  const { title, author, url, likes = 0 } = request.body;
+blogsRouter.post("/", userExtractor, async (request, response, next) => {
+  const { title, url, likes = 0 } = request.body;
+  const { userId } = request;
 
-  if (!title || !author || !url)
+  if (!title || !url)
     return response.status(400).send({
-      error: "title, author and url fields are required"
+      error: "title and url fields are required"
     });
 
   try {
-    const user = await User.findById(author);
+    const user = await User.findById(userId);
     if (!user) return response.status(404).end();
 
     const blog = new Blog({
@@ -60,10 +62,17 @@ blogsRouter.post("/", async (request, response, next) => {
   }
 });
 
-blogsRouter.delete("/:id", async (request, response, next) => {
+blogsRouter.delete("/:id", userExtractor, async (request, response, next) => {
   const { id } = request.params;
+  const { userId } = request;
 
   try {
+    const blog = await Blog.findById(id);
+
+    if (blog && blog.author.toString() !== userId.toString()) {
+      return response.status(401).end();
+    }
+
     const deletedBlog = await Blog.findByIdAndDelete(id);
     return deletedBlog ? response.status(204).end() : response.status(404).end();
   } catch (error) {
@@ -71,7 +80,7 @@ blogsRouter.delete("/:id", async (request, response, next) => {
   }
 });
 
-blogsRouter.put("/:id", async (request, response, next) => {
+blogsRouter.put("/:id", userExtractor, async (request, response, next) => {
   const { id } = request.params;
   const { title, url, likes } = request.body;
